% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.map.R
\name{as.map}
\alias{as.map}
\title{Convert a map from a 3 columns data.frame to a map object}
\usage{
as.map(x)
}
\arguments{
\item{x}{a 3 columns data.frame with marker names, chromosome, and position}
}
\description{
Convert a map from a 3 columns data.frame to a map object
}
\examples{
data(sorgho_p118)
genonmv2<-as.map(genomv2.df)
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

