% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cM.pos}
\alias{cM.pos}
\title{Compute the position in cM from a position in bp}
\usage{
cM.pos(a, conv)
}
\arguments{
\item{a}{a vector of length 2 with chromosome and position in bp as elements}

\item{a}{conv object as constructed by make.conv}
}
\description{
Compute the position in cM from a position in bp
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

