% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cMconvertR}
\alias{cMconvertR}
\title{A wrapper for the cM.pos function}
\usage{
cMconvertR(map, conv, named.chr = F)
}
\arguments{
\item{a}{map}

\item{a}{conv object as constructed by make.conv}
}
\description{
A wrapper for the cM.pos function
}
\examples{
data(sorgho_p118)
genomv2<-as.map(genomv2.df)
p118.conv<-make.conv(mapp = genomv2,mapg = p118.genet, exclude = "SB_06_045")
#create a new dummy map
newmap<-as.map(data.frame(mk=paste("m",1:1000,sep=""),chr=c(rep(1,500),rep(2,500)),Pos=round(c(sort(runif(n=500,min=1,max=1e8)),sort(runif(n=500,min=1,max=1e8))),0)))
newmap.genet<-cMconvertR(newmap,p118.conv)
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

