% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new.pos.R
\name{new.pos}
\alias{new.pos}
\title{Compute the position on map2 from a position on map1}
\usage{
new.pos(a, conv)
}
\arguments{
\item{a}{a vector of length 2 with chromosome and position as elements}

\item{a}{conversion list as constructed by make.conv}
}
\description{
Compute the position on map2 from a position on map1
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

