% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ph.pos.R
\name{ph.pos}
\alias{ph.pos}
\title{Compute the position in bp from a position in cM}
\usage{
ph.pos(a, conv)
}
\arguments{
\item{a}{a vector of length 2 with chromosome and position in cM as elements}

\item{a}{conv object as constructed by make.conv}
}
\description{
Compute the position in bp from a position in cM
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

