% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziplinR.R
\name{ziplinR}
\alias{ziplinR}
\title{The main function to project positions from one map to another}
\usage{
ziplinR(map, map1, map2, named.chr = T, map2tomap1 = F, shift.extrem = T,
  exclude = NULL)
}
\arguments{
\item{map}{a map containing the positions to be projected. The map format is the same as the one used by Rqtl, ie a list of vectors of positions}

\item{map1}{the first reference map}

\item{map2}{the second reference map}

\item{map2tomap1}{if TRUE the positions of map will be projected from map2 to map1. Default is FALSE}

\item{shift.extrem}{if TRUE (default) the new map will be shifted in case some positions are projected to negative values}

\item{exclude}{a vector of marker names to exclude from references, typically showing discrepancies}
}
\value{
a map of new projected positions
}
\description{
The main function to project positions from one map to another
}
\examples{
data(sorgho_p118)
genomv2<-as.map(genomv2.df)
#create a new dummy map
newmap<-as.map(data.frame(mk=paste("m",1:1000,sep=""),chr=c(rep(1,500),rep(2,500)),Pos=round(c(sort(runif(n=500,min=1,max=1e8)),sort(runif(n=500,min=1,max=1e8))),0)))
newmap.genet<-ziplinR(map=newmap,map1=genomv2,map2=p118.genet,exclude = "SB_06_045",shift.extrem = T)
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

