% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.conv.R
\name{make.conv}
\alias{make.conv}
\title{Constuctor for a conversion list}
\usage{
make.conv(map1, map2, exclude = NULL)
}
\arguments{
\item{map1}{}

\item{map2}{}

\item{exclude}{}
}
\description{
Constuctor for a conversion list
}
\examples{
data(sorgho_p118)
genomv2<-as.map(genomv2.df)
p118.conv<-make.conv(map1 = genomv2,map2 = p118.genet, exclude = "SB_06_045")
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

