% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.conv.R
\name{make.conv}
\alias{make.conv}
\title{Constuctor for a conv object}
\usage{
make.conv(mapp, mapg, exclude = NULL)
}
\arguments{
\item{mapp}{}

\item{mapg}{}

\item{exclude}{}
}
\description{
Constuctor for a conv object
}
\examples{
data(sorgho_p118)
genomv2<-as.map(genomv2.df)
p118.conv<-make.conv(mapp = genomv2,mapg = p118.genet, exclude = "SB_06_045")
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

